Fixed miniLockLib.ID.decode(...) to work predictably in Safari.
In Safari, Uint8Array::subarray returns an instance where `buffer.byteLength` corresponds to the original Uint8Array, instead of a `buffer.byteLength` that corresponds with the new subarray. This is not a huge problem, except when we want to compare a subarray using an assertion like deepEqual.

To keep the test assertions working predictably, miniLockLib.ID.decode() constrcuts a new Uint8Array to accept the first 32 bytes of the decoded ID to ensure the buffer.byteLength is defined properly in Safari.