Support file format version 2 with type and time attributes.
Version 1 remains the default at this time.

miniLockLib.encrypt accepts version, type and time params.

In version 1 the first 256 bytes of ciphertext are reserved for the
file name. In version 2 the first 408 bytes of ciphertext are reserved
for the file name, media type and encrypt time. Version 2 files are 152
bytes longer than Version 1 files.